Move header components into directory
This code organization helps us consolidate all the sub-components of
the header into the same directory. The index.js file redirects the
import of the Header component to the header.js file contained within.
