+++ src/components/header/index.js
@@ -1,0 +1,2 @@
+export * from './header';
+export { default } from './header';
